Adjusted header/footer padding, added title section to sidebar
Added mediaqueries to make the mobile/tablet footer a bit prettier, pushed the right and left edges of the navigation bar off the sides to reflect the margins on the title/page content.

Additionally, added a "Your itinerary" title to the sidebar. This helps users make sense of the layout and also gives us a place to stick the "collapse sidebar" button (either to the right or left of the title).