Redesigned home page.
This is a big change- I reworked the home page that a non-user sees when accessing LATE to be a lot nicer-looking and responsive. Tell me if there's any issues.

Additionally, I added a link to the about page in the footer as until now it had become isolated on the site.