Titled all frontend vue files
Not all files are in the right place and it can be confusing to try and tell what you're working on. I added titles and some categories at the top of all the vue files which makes the frontend easier to work on and paves the way for a potential reorganization of files.